requirements.txt
attrs==17.4.0
botocore==1.12.14
certifi==2018.8.24
chalice==1.6.0
chardet==3.0.4
click==6.7
Deprecated==1.2.3
docutils==0.14
enum-compat==0.0.2
idna==2.7
jmespath==0.9.3
PyGithub==1.43.2
PyJWT==1.6.4
python-dateutil==2.7.3
requests==2.19.1
six==1.11.0
typing==3.6.4
urllib3==1.23
